Add cross platform signal handler for stopping work. Some improvements performance

+++ src/domain/DTOs/console_args.hpp
@@ -10,7 +10,7 @@ struct ConsoleArgs {
     const std::filesystem::path passwordsPath;
     const std::filesystem::path proxiesPath;
     const std::string outputPath;
-    unsigned short threadsCount;
+    int threadsCount;
 };
 
 } // namespace domain::dtos
